deepfake_webapp/requirements.txt
flask
torch
torchvision
Pillow
opencv-python
numpy
dlib-bin
exifread
matplotlib
flask-cors
